Product List Scraping - Chewy.com/v2/requirements.txt
tqdm
beautifulsoup4
requests
lxml
numpy
pandas
mysql-connector-python